requirements.txt
Django==1.6
django-excel-response==1.0
django-phonenumber-field==0.6
django-tables2==0.15.0
gnureadline==6.3.3
ipdb==0.8
ipython==2.3.1
phonenumbers==7.0.1
six==1.8.0
wsgiref==0.1.2
xlwt==0.7.5